fix(unet_constructing): fix the error of error data type;

+++ unet_constructing/tf_impl/original.py
@@ -64,13 +64,13 @@ class UNet:
         # layer 1
         with tf.name_scope('layer_1'):
             # conv_1
-            self.w[1] = self.init_w(shape=[3, 3, FLAGS.input_shape[2], 64 / FLAGS.divisor], name='kernel',
+            self.w[1] = self.init_w(shape=[3, 3, FLAGS.input_shape[2], int(64 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv1_1")
             conv_1_result = tf.nn.conv2d(input=normed_batch, filter=self.w[1], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv1_1')
             relu_1_result = tf.nn.relu(conv_1_result, name='relu_1')
             # conv_2
-            self.w[2] = self.init_w(shape=[3, 3, 64 / FLAGS.divisor, 64 / FLAGS.divisor], name='kernel',
+            self.w[2] = self.init_w(shape=[3, 3, int(64 / FLAGS.divisor), int(64 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv1_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[2], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv1_2')
@@ -83,13 +83,13 @@ class UNet:
         # layer 2
         with tf.name_scope('layer_2'):
             # conv_1
-            self.w[3] = self.init_w(shape=[3, 3, 64 / FLAGS.divisor, 128 / FLAGS.divisor], name='kernel',
+            self.w[3] = self.init_w(shape=[3, 3, int(64 / FLAGS.divisor), int(128 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv2_1")
             conv_1_result = tf.nn.conv2d(input=maxpool_result, filter=self.w[3], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv2_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[4] = self.init_w(shape=[3, 3, 128 / FLAGS.divisor, 128 / FLAGS.divisor], name='kernel',
+            self.w[4] = self.init_w(shape=[3, 3, int(128 / FLAGS.divisor), int(128 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv2_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[4], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv2_2')
@@ -102,13 +102,13 @@ class UNet:
         # layer 3
         with tf.name_scope('layer_3'):
             # conv_1
-            self.w[5] = self.init_w(shape=[3, 3, 128 / FLAGS.divisor, 256 / FLAGS.divisor], name='kernel',
+            self.w[5] = self.init_w(shape=[3, 3, int(128 / FLAGS.divisor), int(256 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv3_1")
             conv_1_result = tf.nn.conv2d(
                 input=maxpool_result, filter=self.w[5], strides=[1, 1, 1, 1], padding='SAME', name='conv3_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[6] = self.init_w(shape=[3, 3, 256 / FLAGS.divisor, 256 / FLAGS.divisor], name='kernel',
+            self.w[6] = self.init_w(shape=[3, 3, int(256 / FLAGS.divisor), int(256 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv3_2")
             conv_2_result = tf.nn.conv2d(
                 input=relu_1_result, filter=self.w[6], strides=[1, 1, 1, 1], padding='SAME', name='conv3_2')
@@ -121,13 +121,13 @@ class UNet:
         # layer 4
         with tf.name_scope('layer_4'):
             # conv_1
-            self.w[7] = self.init_w(shape=[3, 3, 256 / FLAGS.divisor, 512 / FLAGS.divisor], name='kernel',
+            self.w[7] = self.init_w(shape=[3, 3, int(256 / FLAGS.divisor), int(512 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv4_1")
             conv_1_result = tf.nn.conv2d(input=maxpool_result, filter=self.w[7], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv4_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[8] = self.init_w(shape=[3, 3, 512 / FLAGS.divisor, 512 / FLAGS.divisor], name='kernel',
+            self.w[8] = self.init_w(shape=[3, 3, int(512 / FLAGS.divisor), int(512 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv4_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[8], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv4_2')
@@ -141,24 +141,24 @@ class UNet:
         # layer 5
         with tf.name_scope('layer_5'):
             # conv_1
-            self.w[9] = self.init_w(shape=[3, 3, 512 / FLAGS.divisor, 1024 / FLAGS.divisor], name='kernel',
+            self.w[9] = self.init_w(shape=[3, 3, int(512 / FLAGS.divisor), int(1024 / FLAGS.divisor)], name='kernel',
                                     scope_name="conv5_1")
             conv_1_result = tf.nn.conv2d(input=maxpool_result, filter=self.w[9], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv5_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[10] = self.init_w(shape=[3, 3, 1024 / FLAGS.divisor, 1024 / FLAGS.divisor], name='kernel',
+            self.w[10] = self.init_w(shape=[3, 3, int(1024 / FLAGS.divisor), int(1024 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv5_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[10], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv5_2')
             relu_2_result = tf.nn.relu(features=conv_2_result, name='relu_2')
 
             # up_sampling
-            self.w[11] = self.init_w(shape=[2, 2, 512 / FLAGS.divisor, 1024 / FLAGS.divisor], name='kernel',
+            self.w[11] = self.init_w(shape=[2, 2, int(512 / FLAGS.divisor), int(1024 / FLAGS.divisor)], name='kernel',
                                      scope_name="Up_Sampling_5")
             result_up = tf.nn.conv2d_transpose(value=relu_2_result, filter=self.w[11], strides=[1, 2, 2, 1],
                                                output_shape=[batch_size, int(relu_2_result.shape[1]) * 2,
-                                                             int(relu_2_result.shape[2]) * 2, 512 / FLAGS.divisor],
+                                                             int(relu_2_result.shape[2]) * 2, int(512 / FLAGS.divisor)],
                                                padding='VALID', name='Up_Sampling_5')
             relu_3_result = tf.nn.relu(features=result_up, name='relu_3')
 
@@ -170,24 +170,24 @@ class UNet:
                 result_from_contracting=self.result_from_contracting[4], result_from_upsampling=relu_3_result)
 
             # conv_1
-            self.w[12] = self.init_w(shape=[3, 3, 1024 / FLAGS.divisor, 512 / FLAGS.divisor], name='kernel',
+            self.w[12] = self.init_w(shape=[3, 3, int(1024 / FLAGS.divisor), int(512 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv6_1")
             conv_1_result = tf.nn.conv2d(input=result_merge, filter=self.w[12], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv6_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[13] = self.init_w(shape=[3, 3, 512 / FLAGS.divisor, 512 / FLAGS.divisor], name='kernel',
+            self.w[13] = self.init_w(shape=[3, 3, int(512 / FLAGS.divisor), int(512 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv6_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[13], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv6_2')
             relu_2_result = tf.nn.relu(features=conv_2_result, name='relu_2')
 
             # up_sampling
-            self.w[14] = self.init_w(shape=[2, 2, 256 / FLAGS.divisor, 512 / FLAGS.divisor], name='kernel',
+            self.w[14] = self.init_w(shape=[2, 2, int(256 / FLAGS.divisor), int(512 / FLAGS.divisor)], name='kernel',
                                      scope_name="Up_Sampling_6")
             result_up = tf.nn.conv2d_transpose(value=relu_2_result, filter=self.w[14], strides=[1, 2, 2, 1],
                                                output_shape=[batch_size, int(relu_2_result.shape[1]) * 2,
-                                                             int(relu_2_result.shape[2]) * 2, 256 / FLAGS.divisor],
+                                                             int(relu_2_result.shape[2]) * 2, int(256 / FLAGS.divisor)],
                                                padding='VALID', name='Up_Sampling_6')
             relu_3_result = tf.nn.relu(features=result_up, name='relu_3')
 
@@ -198,24 +198,24 @@ class UNet:
                 result_from_contracting=self.result_from_contracting[3], result_from_upsampling=relu_3_result)
 
             # conv_1
-            self.w[15] = self.init_w(shape=[3, 3, 512 / FLAGS.divisor, 256 / FLAGS.divisor], name='kernel',
+            self.w[15] = self.init_w(shape=[3, 3, int(512 / FLAGS.divisor), int(256 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv7_1")
             conv_1_result = tf.nn.conv2d(input=result_merge, filter=self.w[15], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv7_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[16] = self.init_w(shape=[3, 3, 256 / FLAGS.divisor, 256 / FLAGS.divisor], name='kernel',
+            self.w[16] = self.init_w(shape=[3, 3, int(256 / FLAGS.divisor), int(256 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv7_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[16], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv7_2')
             relu_2_result = tf.nn.relu(features=conv_2_result, name='relu_2')
 
             # up_sampling
-            self.w[17] = self.init_w(shape=[2, 2, 128 / FLAGS.divisor, 256 / FLAGS.divisor], name='kernel',
+            self.w[17] = self.init_w(shape=[2, 2, int(128 / FLAGS.divisor), int(256 / FLAGS.divisor)], name='kernel',
                                      scope_name="Up_Sampling_7")
             result_up = tf.nn.conv2d_transpose(value=relu_2_result, filter=self.w[17], strides=[1, 2, 2, 1],
                                                output_shape=[batch_size, int(relu_2_result.shape[1]) * 2,
-                                                             int(relu_2_result.shape[2]) * 2, 128 / FLAGS.divisor],
+                                                             int(relu_2_result.shape[2]) * 2, int(128 / FLAGS.divisor)],
                                                padding='VALID', name='Up_Sampling_7')
             relu_3_result = tf.nn.relu(features=result_up, name='relu_3')
 
@@ -226,13 +226,13 @@ class UNet:
                 result_from_contracting=self.result_from_contracting[2], result_from_upsampling=relu_3_result)
 
             # conv_1
-            self.w[18] = self.init_w(shape=[3, 3, 256 / FLAGS.divisor, 128 / FLAGS.divisor], name='kernel',
+            self.w[18] = self.init_w(shape=[3, 3, int(256 / FLAGS.divisor), int(128 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv8_1")
             conv_1_result = tf.nn.conv2d(input=result_merge, filter=self.w[18], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv8_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[19] = self.init_w(shape=[3, 3, 128 / FLAGS.divisor, 128 / FLAGS.divisor], name='kernel',
+            self.w[19] = self.init_w(shape=[3, 3, int(128 / FLAGS.divisor), int(128 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv8_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[19], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv8_2')
@@ -242,7 +242,7 @@ class UNet:
             self.w[20] = self.init_w(shape=[2, 2, 64, 128], name='kernel', scope_name="Up_Sampling_8")
             result_up = tf.nn.conv2d_transpose(value=relu_2_result, filter=self.w[20], strides=[1, 2, 2, 1],
                                                output_shape=[batch_size, int(relu_2_result.shape[1]) * 2,
-                                                             int(relu_2_result.shape[2]) * 2, 64 / FLAGS.divisor],
+                                                             int(relu_2_result.shape[2]) * 2, int(64 / FLAGS.divisor)],
                                                padding='VALID', name='Up_Sampling_8')
             relu_3_result = tf.nn.relu(features=result_up, name='relu_3')
 
@@ -253,19 +253,19 @@ class UNet:
                 result_from_contracting=self.result_from_contracting[1], result_from_upsampling=relu_3_result)
 
             # conv_1
-            self.w[21] = self.init_w(shape=[3, 3, 128 / FLAGS.divisor, 64 / FLAGS.divisor], name='kernel',
+            self.w[21] = self.init_w(shape=[3, 3, int(128 / FLAGS.divisor), int(64 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv9_1")
             conv_1_result = tf.nn.conv2d(input=result_merge, filter=self.w[21], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv9_1')
             relu_1_result = tf.nn.relu(features=conv_1_result, name='relu_1')
             # conv_2
-            self.w[22] = self.init_w(shape=[3, 3, 64 / FLAGS.divisor, 64 / FLAGS.divisor], name='kernel',
+            self.w[22] = self.init_w(shape=[3, 3, int(64 / FLAGS.divisor), int(64 / FLAGS.divisor)], name='kernel',
                                      scope_name="conv9_2")
             conv_2_result = tf.nn.conv2d(input=relu_1_result, filter=self.w[22], strides=[1, 1, 1, 1],
                                          padding='SAME', name='conv9_2')
 
             relu_2_result = tf.nn.relu(features=conv_2_result, name='relu_2')
-            self.w[23] = self.init_w(shape=[1, 1, 64 / FLAGS.divisor, 3], name='kernel', scope_name="conv9_3")
+            self.w[23] = self.init_w(shape=[1, 1, int(64 / FLAGS.divisor), 3], name='kernel', scope_name="conv9_3")
             result_conv_3 = tf.nn.conv2d(input=relu_2_result, filter=self.w[23], strides=[1, 1, 1, 1],
                                          padding='VALID', name='conv9_3')
